Fixes serial connection to inverter
The timeout property in a serial context means how long
a read can be ongoing trying to fill the supplied buffer.

The default value of 1 second creates a situation where the
simulated modbus slave appears to be nonresponsive from the
SolarEdge inverter's perspective, as it issues a read request
and for up to a second, the slave does not respond. The WattNode
physical device is specified to respond within 200 ms. Its
default setting with regards to parity is N1, i.e. no parity, 1
stop bit.

Allow fractional seconds as timeout and lower the serial timeout
to 0.1 s, which risks more incomplete RTU frames to be returned
from the serial library, but this is normal and handled by the
pymodbus library.